requirements.txt
Django>=5.0,<6.1
psycopg2-binary>=2.9
django-htmx>=1.17
python-dotenv>=1.0
whitenoise>=6.6
gunicorn>=21.2
Pillow>=10.2
colorlog>=6.8
ruff>=0.1.0
djlint>=1.34.0
